% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{edatTimingFiles}
\alias{edatTimingFiles}
\title{edatTimingFiles}
\usage{
edatTimingFiles(
  datfile,
  datpath,
  savedir,
  pid,
  session,
  skiprows = 1,
  sheetname = NA,
  prefix = "NN",
  timingfile_format = "default",
  condition_labels = "default",
  figure = NULL,
  noreturn = TRUE
)
}
\arguments{
\item{datfile}{Name of the edat text file}

\item{datpath}{path to the datfile}

\item{savedir}{path to the folder where timing files will be saved}

\item{pid}{Participant ID}

\item{session}{Scan Session ID}

\item{skiprows}{(optional) number of rows to skip when importing the edatfile. Default is '1', meaning that column names start on row 2.}

\item{sheetname}{(optional) If datfile is an excel sheet instead of a text file, the name of the Sheet to import}

\item{prefix}{(optional) a parameter that can be used in 'timingfile_format'. Defaults to 'NN'}

\item{timingfile_format}{(optional) A string containing R code to evaluate to create the filenames of the timing files. Default is:
"sprintf('\%s\%s_\%s_\%s.txt', prefix, pid, session, condition)"}

\item{condition_labels}{(optional) A list of condition-label pairs. The keys are the conditions in the edat file. The values are the strings that get used to create the timingfile filenames.}

\item{figure}{(optional) A full path+filename to save a timing file figure. Defaults to NULL (i.e., do not make a figure)}

\item{noreturn}{(optional) if TRUE (default), the function will return NULL instead of the timing data frame. To return the timing data, set to FALSE.
#'}
}
\description{
Read an edat text file and create timing files based on condition names
for use with afni preprocessing. Returns a data frame of onset times for condition and run.
}
\details{
Usage:
edatTimingFiles('MID_NN-RCT-2002_T1.txt','data','output', 1001, 1)
}
