% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{edatTimingFiles}
\alias{edatTimingFiles}
\title{edatTimingFiles}
\usage{
edatTimingFiles(
  datfile,
  datpath,
  savedir,
  pid,
  session,
  skiprows = "auto",
  sheetname = NA,
  prefix = "NN",
  timingfile_format = "default",
  condition_labels = "default",
  figure = NULL,
  noreturn = TRUE,
  proc_method = "MID",
  export_format = "afni"
)
}
\arguments{
\item{datfile}{Name of the edat text file}

\item{datpath}{path to the datfile}

\item{savedir}{path to the folder where timing files will be saved}

\item{pid}{Participant ID}

\item{session}{Scan Session ID}

\item{skiprows}{(optional) number of rows to skip when importing the edatfile. Default is 'auto': If the first line contains '.edat3' it will skip that line and start on 2, otherwise it will start on the first line.}

\item{sheetname}{(optional) If datfile is an excel sheet instead of a text file, the name of the Sheet to import}

\item{prefix}{(optional) a parameter that can be used in 'timingfile_format'. Defaults to 'NN'}

\item{timingfile_format}{(optional) A string containing R code to evaluate to create the filenames of the timing files. Default is:
"sprintf('\%s\%s_\%s_\%s.txt', prefix, pid, session, condition)"}

\item{condition_labels}{(optional) A list of condition-label pairs. The keys are the conditions in the edat file. The values are the strings that get used to create the timingfile filenames.}

\item{figure}{(optional) A full path+filename to save a timing file figure. Defaults to NULL (i.e., do not make a figure)}

\item{noreturn}{(optional) if TRUE (default), the function will return NULL instead of the timing data frame. To return the timing data, set to FALSE.}

\item{proc_method}{(optional) Determines which internal function is used to detect block/trial onsets and conditions. Defaults to 'MID', monetary incentive delay. Also supports 'nback'.}

\item{export_format}{(optional) Defaults to 'AFNI', and creates a separate timing file for each condition containing event onsets. Also supports 'fmriprep' which outputs a BIDS-compliant events.tsv file.}
}
\description{
Read an edat text file and create timing files based on condition names
for use with afni preprocessing. Returns a data frame of onset times for condition and run.
Input can either be a raw eprime text file (i.e., first line "*** Header Start ***"), or an exported eprime text file (i.e., a tab delimited table -- a processed version of the raw eprime text file).
}
\details{
Example function calls:

edatTimingFiles('MID_NN-RCT-2002_T1.txt','data','output', 1001, 1)

edatTimingFiles('NBACK.txt','data','output','001','BL', timingfile_format = "sprintf('sub-\%s_ses-\%s_task-NBack_events.tsv', pid, session)", proc_method = 'nback', export_format = 'fmriprep')
}
